Prometheus and label changing ain't easy.
The default instrumentation, expects labels code and method. We don't
use code labels, instead we do use status_code.

A process of relabeling is possible, it does expect extra work on
implementing changes on dashboard before all services are migrated to a
new service code.

It does not bring a lot of benefit, but generates extra work. I will
stick to labels we already have and WE WILL NOT BE BLOCK ON RELEASING of
rpc-gateway.