Remove a concept of rerouting
An option AllowedNumberOfReroutes provides a way to define a number
of reroutes among different node providers, e.g. when you define 3 node
providers, you may set AllowedNumberOfReroutes to 2 and the last node
provider will never be reached. You can call it a misconfiguration when
first two node providers are experiencing an outage.

When you have a backend with a single node provider, a number of
rereoutes does not make sense, because this where number of retries per
backend should be used.

When you define a backend of multiple node providers, your
intention is to allow reroutes among them, otherwise what's the point of
multi node provider backend?